requirements.txt
py-postgresql==1.2.1
telebot==3.5.1